finish 29 level without big task

+++ 3.JavaMultithreading/src/com/javarush/task/task29/task2906/Solution.java
@@ -15,6 +15,6 @@ public class Solution {
     }
 
     public static Integer getValue(boolean first, boolean second) {
-        return first ? 100 : second ? 200 : null;
+        return first ? (Integer) 100 : second ? 200 : null;
     }
 }
